requirements.txt
numpy
pandas
torch
torchvision
matplotlib
scikit-learn
opencv-python
imageio
Pillow
streamlit
tqdm
